+++ ucm/shared/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(vendor)
+add_subdirectory(infra)
 add_subdirectory(trans)
 add_subdirectory(test)
