+++ ucm/shared/infra/logger/logger.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_LOGGER_H
-#define UNIFIEDCACHE_LOGGER_H
+#ifndef UNIFIEDCACHE_INFRA_LOGGER_H
+#define UNIFIEDCACHE_INFRA_LOGGER_H
 
 #include <cstddef>
 #include <fmt/format.h>
