+++ ucm/shared/infra/logger/spdlog/spdlog_logger.cc
