+++ ucm/shared/infra/template/hashset.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_HASHSET_H
-#define UNIFIEDCACHE_HASHSET_H
+#ifndef UNIFIEDCACHE_INFRA_HASHSET_H
+#define UNIFIEDCACHE_INFRA_HASHSET_H
 
 #include <array>
 #include <atomic>
