+++ ucm/shared/infra/template/singleton.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_SINGLETON_H
-#define UNIFIEDCACHE_SINGLETON_H
+#ifndef UNIFIEDCACHE_INFRA_SINGLETON_H
+#define UNIFIEDCACHE_INFRA_SINGLETON_H
 
 namespace UC {
