+++ ucm/shared/infra/thread/index_pool.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_INDEX_POOL_H
-#define UNIFIEDCACHE_INDEX_POOL_H
+#ifndef UNIFIEDCACHE_INFRA_INDEX_POOL_H
+#define UNIFIEDCACHE_INFRA_INDEX_POOL_H
 
 #include <atomic>
 #include <cstddef>
