+++ ucm/shared/infra/thread/thread_pool.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_THREAD_POOL_H
-#define UNIFIEDCACHE_THREAD_POOL_H
+#ifndef UNIFIEDCACHE_INFRA_THREAD_POOL_H
+#define UNIFIEDCACHE_INFRA_THREAD_POOL_H
 
 #include <condition_variable>
 #include <functional>
