+++ ucm/shared/infra/time/stopwatch.h
@@ -21,8 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * */
-#ifndef UNIFIEDCACHE_STOPWATCH_H
-#define UNIFIEDCACHE_STOPWATCH_H
+#ifndef UNIFIEDCACHE_INFRA_STOPWATCH_H
+#define UNIFIEDCACHE_INFRA_STOPWATCH_H
 
 #include <chrono>
