+++ ucm/shared/trans/CMakeLists.txt
@@ -8,6 +8,7 @@ if(RUNTIME_ENVIRONMENT STREQUAL "simu")
     add_subdirectory(simu)
 endif()
 target_include_directories(trans PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/..)
+target_link_libraries(trans PUBLIC infra_status)
 
 file(GLOB_RECURSE UCMTRANS_CPY_SOURCE_FILES "./cpy/*.cc")
 pybind11_add_module(ucmtrans ${UCMTRANS_CPY_SOURCE_FILES})
