+++ ucm/shared/trans/stream.h
@@ -25,7 +25,7 @@
 #define UNIFIEDCACHE_TRANS_STREAM_H
 
 #include <functional>
-#include "status.h"
+#include "status/status.h"
 
 namespace UC::Trans {
